+++ src/symbols/Add/Add.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import classNames from 'classnames';
-import { motion } from 'framer-motion';
+import { motion } from 'motion/react';
 import css from './Add.module.css';
 
 export interface AddProps {
@@ -14,16 +14,7 @@ export interface AddProps {
   onClick?: (event: React.MouseEvent<SVGGElement, MouseEvent>) => void;
 }
 
-export const Add: FC<Partial<AddProps>> = ({
-  x,
-  y,
-  className,
-  size = 15,
-  hidden = true,
-  onEnter = () => undefined,
-  onLeave = () => undefined,
-  onClick = () => undefined
-}) => {
+export const Add: FC<Partial<AddProps>> = ({ x, y, className, size = 15, hidden = true, onEnter = () => undefined, onLeave = () => undefined, onClick = () => undefined }) => {
   if (hidden) {
     return null;
   }
@@ -33,13 +24,7 @@ export const Add: FC<Partial<AddProps>> = ({
   const translateY = y - half;
 
   return (
-    <motion.g
-      className={classNames(className, css.container)}
-      initial={{ scale: 0, opacity: 0, translateX, translateY }}
-      animate={{ scale: 1, opacity: 1, translateX, translateY }}
-      whileHover={{ scale: 1.2 }}
-      whileTap={{ scale: 0.8 }}
-    >
+    <motion.g className={classNames(className, css.container)} initial={{ scale: 0, opacity: 0, translateX, translateY }} animate={{ scale: 1, opacity: 1, translateX, translateY }} whileHover={{ scale: 1.2 }} whileTap={{ scale: 0.8 }}>
       <rect
         height={size * 2}
         width={size * 2}
@@ -53,22 +38,8 @@ export const Add: FC<Partial<AddProps>> = ({
         onMouseLeave={onLeave}
       />
       <rect height={size} width={size} className={css.rect} />
-      <line
-        x1="2"
-        x2={size - 2}
-        y1={half}
-        y2={half}
-        className={css.plus}
-        strokeWidth="1"
-      />
-      <line
-        x1={half}
-        x2={half}
-        y1="2"
-        y2={size - 2}
-        className={css.plus}
-        strokeWidth="1"
-      />
+      <line x1="2" x2={size - 2} y1={half} y2={half} className={css.plus} strokeWidth="1" />
+      <line x1={half} x2={half} y1="2" y2={size - 2} className={css.plus} strokeWidth="1" />
     </motion.g>
   );
 };